backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
openai==1.14.0
pydantic==2.6.4
python-dotenv==1.0.1
pydantic-settings==2.2.1
